CPU should print correct numbers for unrecognised bytecodes

+++ src/cpu/cpu.cpp
@@ -405,7 +405,7 @@ byte* CPU::dispatch(byte* addr) {
             break;
         default:
             ostringstream error;
-            error << "unrecognised instruction (bytecode value: " << *((int*)bytecode) << ")";
+            error << "unrecognised instruction (bytecode value: " << *((byte*)bytecode) << ")";
             throw error.str().c_str();
     }
     return addr;
